==> scripts/01_install_packages.R <==
## This script installs all packages necessary for data cleaning, analysis, and visualization

install.packages("tidyverse") # used for dataframe manipulation (primarily dplyr) and visualizations (ggplot2)
install.packages("lintr")
install.packages("gridExtra") # used for some plotting
install.packages("grid")
install.packages("plotly") # used for interactive heatmaps
install.packages("VennDiagram") # used for erosion diagram
install.packages("RColorBrewer") # used for colors
